pyproject.toml
[build-system]
requires = ["uv_build==0.7.19"]
build-backend = "uv_build"

[project]
name = "h_rag"
version = "0.1"
readme = "README.md"
requires-python = "==3.13.12"
dependencies = [
    "chromadb>=1.5.0",
    "dotenv>=0.9.9",
    "google-genai>=1.63.0",
    "loguru>=0.7.3",
    "ollama>=0.6.1",
    "pymupdf>=1.27.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.54.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "ruff==0.15.1",
    "pytest==8.4.1",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = []

[tool.ruff]
include = ["src/**.py"]
exclude = ["src/**/__init__.py"]
line-length = 100

[tool.ruff.lint]
select = ["D", "F401", "I"]
ignore = ["D107"]


[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.ruf.format]
docstring-code-format = true
line-ending = "lf"

[tool.mypy]
mypy_path = ["src"]
